Split main into package files
